abstracted min and max value aggregators
